tq/pyproject.toml
[project]
name = "tq"
version = "0.0.1"
description = "Helper package for Turquoise Health research"
keywords = [
  "healthcare",
  "data science",
  "data analysis",
  "healthcare pricing"
]
authors = [
  {name = "Dan Snow", email = "dan.snow@turquoise.health"}
]
classifiers = [
  "Development Status :: 1 - Planning",
  "Intended Audience :: Developers",
  "Intended Audience :: Science/Research",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13"
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
  "GitPython>=3.1.0",
  "python-dotenv>=1.1.0",
  "setuptools>=75.0.0",
  "trino>=0.333.0"
]

[project.urls]
repository = "https://github.com/turquoisehealth/research/tq"

[project.optional-dependencies]
dev = [
  "pre-commit>=4.0.1",
  "pytest>=7.3.0",
  "pytest-cov>=4.1.0"
]

# Packaging and build tools
[build-system]
requires = ["setuptools>=75.0.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
include-package-data = true

[tool.setuptools.package-data]
"*" = ["*.parquet"]

[tool.uv]
package = true

# CI and test tools
[tool.pytest.ini_options]
minversion = "7.0.0"
addopts = "-v --cache-clear -rf --cov=tq --cov-report=term --cov-report=html --maxfail=10"
console_output_style = "count"
python_classes = "Test"
python_files = "test_*.py"
python_functions = "test_*"
testpaths = "tq/tests"

[tool.tox]
legacy_tox_ini = """
[tox]
min_version = 4.0
envlist =
    py{310, 311, 312, 313}

[gh-actions]
python =
    3.10: py310
    3.11: py311
    3.12: py312
    3.13: py313

[testenv]
extras = dev
commands = pytest
passenv =
    UV_CACHE_DIR
    PYTHONUNBUFFERED

[testenv:py{310, 311, 312, 313}]
uv_resolution = highest
"""

# Linting and type checking tool configs inherited from the root pyproject.toml
[tool.pyright]
extends = "../pyproject.toml"

[tool.ruff]
extend = "../pyproject.toml"
